change type of race_distance, add tests, add methods to parse text files

+++ scripts/race_results.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS race_results (
     race_type TEXT NOT NULL DEFAULT 'personal',
     race_date DATE,
     race_name TEXT,
-    race_distance DOUBLE PRECISION NOT NULL,
+    race_distance INTEGER NOT NULL,
     race_time DOUBLE PRECISION NOT NULL,
     race_flag BOOLEAN NOT NULL DEFAULT FALSE
 );
